server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.7
cerebras-cloud-sdk
chromadb
